# perf: remove reflect from BackendRefContext