--- a/internal/gatewayapi/listener.go
+++ b/internal/gatewayapi/listener.go
@@ -819,7 +819,8 @@ func (t *Translator) processBackendRefs(name string, backendCluster egv1a1.Backe
 		return nil, nil, err
 	}
 	result := make([]*ir.DestinationSetting, 0, len(backendCluster.BackendRefs))
-	for _, ref := range backendCluster.BackendRefs {
+	for i := range backendCluster.BackendRefs {
+		ref := &backendCluster.BackendRefs[i]
 		ns := NamespaceDerefOr(ref.Namespace, namespace)
 		kind := KindDerefOr(ref.Kind, resource.KindService)
 		switch kind {